+++ src/shared/components/toolbar/scientific-toolbar.ts
