+++ src/shared/components/toolbar/stories/scientific-toolbar.stories.ts
